Removing the existing floating package and replacing it with the new one. Still some bugs and docking is completely missing.
