+++ docking-api/src/ModernDocking/floating/DockedTabbedPanelFloatListener.java
@@ -1,0 +1,64 @@
+/*
+Copyright (c) 2024 Andrew Auclair
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+ */
+package ModernDocking.floating;
+
+import ModernDocking.api.DockingAPI;
+import ModernDocking.internal.DockableWrapper;
+import ModernDocking.internal.DockedTabbedPanel;
+
+import javax.swing.*;
+import java.awt.*;
+import java.util.ArrayList;
+import java.util.List;
+
+public class DockedTabbedPanelFloatListener extends FloatListener {
+    private final DockingAPI docking;
+    private final DockedTabbedPanel tabs;
+
+    public DockedTabbedPanelFloatListener(DockingAPI docking, DockedTabbedPanel tabs, JComponent dragComponent) {
+        super(docking, tabs, dragComponent);
+
+        this.docking = docking;
+        this.tabs = tabs;
+    }
+
+    @Override
+    protected Window getOriginalWindow() {
+        return SwingUtilities.windowForComponent(tabs);
+    }
+
+    @Override
+    protected void undock() {
+        List<DockableWrapper> wrappers = new ArrayList<>(tabs.getDockables());
+
+        for (DockableWrapper wrapper : wrappers) {
+            docking.undock(wrapper.getDockable());
+        }
+    }
+
+    @Override
+    protected JFrame createFloatingFrame() {
+        List<DockableWrapper> wrappers = new ArrayList<>(tabs.getDockables());
+
+        return new TempFloatingFrame(wrappers, tabs.getSelectedTabIndex(), tabs, tabs.getSize());
+    }
+}
