--- docking-api/src/ModernDocking/floating/DockingEmptyOverlay.java
@@ -1,10 +1,0 @@
-package ModernDocking.floating;
-
-import java.awt.*;
-
-public class DockingEmptyOverlay extends DockingOverlayInterface {
-    @Override
-    public void paint(Graphics g) {
-        super.paint(g);
-    }
-}
