+++ docking-api/src/ModernDocking/floating/DockingHandle.java
@@ -46,6 +46,8 @@ public class DockingHandle extends JLabel {
 	private final boolean isRoot;
 	private final boolean isPin;
 
+	private boolean mouseOver = false;
+
 	/**
 	 * Create a new DockingHandle
 	 *
@@ -70,6 +72,14 @@ public class DockingHandle extends JLabel {
 		// set the bounds (we're not in a layout manager) and make sure this handle isn't visible
 		setBounds(0, 0, HANDLE_ICON_SIZE, HANDLE_ICON_SIZE);
 		setVisible(false);
+	}
+
+	public void mouseMoved(Point mousePosition) {
+		mouseOver = getBounds().contains(mousePosition);
+	}
+
+	public boolean isMouseOver() {
+		return mouseOver;
 	}
 
 	/**
@@ -106,10 +116,11 @@ public class DockingHandle extends JLabel {
 	 * @param g2 used to draw the dashed lines on top
 	 * @param mouseOver is the mouse over this handle?
 	 */
-	public void paintHandle(Graphics g, Graphics2D g2, boolean mouseOver) {
+	public void paintHandle(Graphics g, Graphics2D g2) {
 		if (!isVisible()) {
 			return;
 		}
+
 		Rectangle bounds = getBounds();
 
 		Color background = DockingSettings.getHandleBackground();
