+++ docking-api/src/ModernDocking/floating/RootDockingHandles.java
@@ -19,7 +19,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-package ModernDocking.floating2;
+package ModernDocking.floating;
 
 import ModernDocking.Dockable;
 import ModernDocking.DockableStyle;
